Moved files away from core and changed gitignore
