--- src/core/index.ts
@@ -1,3 +1,0 @@
-export * from "./factories";
-export * from "./errors";
-export * from "./validation";
