+++ src/factories.ts
@@ -1,5 +1,5 @@
-import { DSNPUserURI } from "../utilities/types";
-import { hash } from "../utilities/hash";
+import { DSNPUserURI } from "./utilities/types";
+import { hash } from "./utilities/hash";
 
 interface ActivityContentBase {
   "@context": "https://www.w3.org/ns/activitystreams";
